:recycle: Rename RepositorySelector.dart to repository_selector.dart
