+++ lib/ui/settings/settings_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:gitdone/ui/settings/widgets/RepositorySelector.dart';
+import 'package:gitdone/ui/settings/widgets/repository_selector.dart';
 
 class SettingsView extends StatelessWidget {
   const SettingsView({super.key});
